# Conflicts:
#	MyApplication/.idea/modules.xml
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/MyApplication.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/activity/HomeActivity.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/activity/LoginActivity.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/activity/SiJiaoDetailActivity.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/activity/SplashActivity.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/bean/Data.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/bean/ForOtherListBean.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/shouhuan/service/BluetoothLeService.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/utils/Constants.java
#	MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/view/CustomLine.java
#	MyApplication/donglan/src/main/res/layout/activity_publish.xml
#	MyApplication/donglan/src/main/res/layout/sijiao_detail.xml
#	MyApplication/sdk/src/main/java/com/tencent/qcloud/sdk/Constant.java
#	MyApplication/settings.gradle
#	MyApplication/tlslibrary/src/main/java/com/tencent/qcloud/tlslibrary/service/TlsBusiness.java

+++ MyApplication/donglan/src/main/java/com/cn/danceland/myapplication/bean/store/storetype/StoreTypeCond.java
@@ -14,8 +14,9 @@ public class StoreTypeCond extends BaseCondition  {
 	private Float face;// 面值
 	private Float giving;// 赠送金额
 	private String remark;// 备注
-	private Byte enable;// 状态
+	private int enable;// 状态
 	private String name;// 储值卡名称
+
 
 	public Long getId() {
 		return id;
@@ -47,10 +48,10 @@ public class StoreTypeCond extends BaseCondition  {
 	public void setRemark(String remark) {
 		this.remark = remark;
 	}
-	public Byte getEnable() {
+	public int getEnable() {
 		return enable;
 	}
-	public void setEnable(Byte enable) {
+	public void setEnable(int enable) {
 		this.enable = enable;
 	}
 	public String getName() {
